+++ lib/trailblazer/activity/dsl/linear/normalizer.rb
@@ -2,6 +2,13 @@ module Trailblazer
   class Activity
     module DSL
       module Linear
+        class KwargsRunner # FIXME: name, location.
+          def self.call(task, ctx, flow_options, circuit_options)
+            task.(ctx, flow_options, circuit_options, **ctx.to_h) # return ctx, flow_options
+          end
+        end
+
+
         # Normalizers are linear activities that process and normalize the options from a specific DSL call,
         # such as `#step` or `#pass`. All defaulting should happen through the normalizer. An invoked
         # normalizer produces an options hash that has to contain an [:adds] key with a ADDS structure usable
@@ -12,21 +19,6 @@ module Trailblazer
         #
         # Most parts of Normalizer are documented: https://trailblazer.to/2.1/docs/internals.html#internals-dsl-normalizer
         module Normalizer
-          # Implements adapter for a callable in a Pipeline that's called with
-          # kwargs and returns a new {ctx}.
-          class TaskAdapter #< Circuit::TaskAdapter
-            def initialize(filter)
-              @filter = filter
-            end
-            def call(wrap_ctx, flow_options, circuit_options)
-              ctx, _ = @filter.(wrap_ctx, **wrap_ctx)
-
-              # If normalizer step returns nil, we use the old ctx.
-              # FIXME: this (or) sucks.
-              return (ctx || wrap_ctx), flow_options # FIXME: signal!
-            end
-          end # TaskAdapter
-
           # Container for all final normalizers of a specific Strategy.
           class Normalizers
             def initialize(**options)
@@ -36,8 +28,12 @@ module Trailblazer
             # Execute the specific normalizer (step, fail, pass) for a particular option set provided
             # by the DSL user. Usually invoked when you call {#step}.
             def call(name, ctx)
+              flow_options = {} # This can be used for tracing, at some point.
+
               normalizer = @normalizers.fetch(name)
-              wrap_ctx, _ = normalizer.(ctx, nil, nil) # TODO: tracing/flow_options?
+
+              wrap_ctx, _ = Pipeline.(normalizer, ctx, flow_options, runner: KwargsRunner) # FIXME: experimental feature from {activity}.
+
               wrap_ctx
             end
           end
@@ -74,19 +70,6 @@ module Trailblazer
             end
           end
 
-          # Wrap user's normalizer task in a {Pipeline::TaskAdapter} so it executes with
-          # convenient kw args.
-          #
-          # Example
-          #   normalizer_task = Normalizer.Task(method(:normalize_id))
-          #
-          #   # will call {normalizer_task} and pass ctx variables as kwargs, as follows
-          #   def normalize_id(ctx, id: false, task:, **)
-          def Task(user_step)
-            # TaskAdapter.for_step(user_step, option: false) # we don't need Option as we don't have ciruit_options here, and no {:exec_context}
-            TaskAdapter.new(user_step)
-          end
-
           # The generic normalizer not tied to `step` or friends.
           def Normalizer(prepend_to_default_outputs: {})
             # Adding steps to the output pipeline means they are only called when there
@@ -96,67 +79,67 @@ module Trailblazer
 
             # Call the prepend_to_outputs pipeline only if {:outputs} is not set (by Subprocess).`
             # too bad we don't have nesting here, yet.
-            defaults_for_outputs = ->(ctx, flow_options, circuit_options) do
-              return [ctx, flow_options] if ctx.key?(:outputs)
+            defaults_for_outputs = ->(ctx, flow_options, circuit_options, outputs: nil, **) do
+              return ctx, flow_options if outputs
 
-              outputs_pipeline.(ctx, flow_options, circuit_options)
+              Pipeline.(outputs_pipeline, ctx, flow_options, circuit_options) # DISCUSS: redundant
             end
 
             Activity::Pipeline(
               {
-                "activity.normalize_step_interface"       => Normalizer.Task(method(:normalize_step_interface)),        # Makes sure {:options} is always a hash.
-                "activity.macro_options_with_symbol_task" => Normalizer.Task(method(:macro_options_with_symbol_task)),  # DISCUSS: we might deprecate {task: :instance_method}
+                "activity.normalize_step_interface"       => method(:normalize_step_interface),        # Makes sure {:options} is always a hash.
+                "activity.macro_options_with_symbol_task" => method(:macro_options_with_symbol_task),  # DISCUSS: we might deprecate {task: :instance_method}
 
-                "activity.merge_library_options"          => Normalizer.Task(method(:merge_library_options)),    # Merge "macro"/user options over library options.
-                "activity.normalize_for_macro"            => Normalizer.Task(method(:merge_user_options)),       # Merge user_options over "macro" options.
-                "activity.normalize_normalizer_options"   => Normalizer.Task(method(:merge_normalizer_options)), # Merge user_options over normalizer_options.
+                "activity.merge_library_options"          => method(:merge_library_options),    # Merge "macro"/user options over library options.
+                "activity.normalize_for_macro"            => method(:merge_user_options),       # Merge user_options over "macro" options.
+                "activity.normalize_normalizer_options"   => method(:merge_normalizer_options), # Merge user_options over normalizer_options.
 
-                "activity.path_helper.forward_block"      => Normalizer.Task(Helper::Path::Normalizer.method(:forward_block_for_path_branch)),     # forward the "global" block
+                "activity.path_helper.forward_block"      => Helper::Path::Normalizer.method(:forward_block_for_path_branch),     # forward the "global" block
 
                 "activity.normalize_context"              => method(:normalize_context),
-                "activity.id_with_inherit_and_replace"    => Normalizer.Task(method(:id_with_inherit_and_replace)),
-                "activity.normalize_id"                   => Normalizer.Task(method(:normalize_id)),
-                "activity.wrap_task_with_step_interface"  => Normalizer.Task(method(:wrap_task_with_step_interface)),
+                "activity.id_with_inherit_and_replace"    => method(:id_with_inherit_and_replace),
+                "activity.normalize_id"                   => method(:normalize_id),
+                "activity.wrap_task_with_step_interface"  => method(:wrap_task_with_step_interface),
 
                 # Nested pipeline:
                 "activity.default_outputs"                => defaults_for_outputs, # only {if :outputs.nil?}
 
-                "inherit.recall_recorded_options"         => Normalizer.Task(Inherit.method(:recall_recorded_options)),
+                "inherit.recall_recorded_options"         => Inherit.method(:recall_recorded_options),
 
-                "activity.sequence_insert"                => Normalizer.Task(method(:normalize_sequence_insert)),
-                "activity.normalize_duplications"         => Normalizer.Task(method(:normalize_duplications)),
+                "activity.sequence_insert"                => method(:normalize_sequence_insert),
+                "activity.normalize_duplications"         => method(:normalize_duplications),
 
-                "activity.path_helper.path_to_track"      => Normalizer.Task(Helper::Path::Normalizer.method(:convert_paths_to_tracks)),
+                "activity.path_helper.path_to_track"      => Helper::Path::Normalizer.method(:convert_paths_to_tracks),
 
 
-                "output_tuples.normalize_output_tuples"           => Normalizer.Task(OutputTuples.method(:normalize_output_tuples)),     # Output(Signal, :semantic) => Id()
-                "output_tuples.remember_custom_output_tuples"     => Normalizer.Task(OutputTuples.method(:remember_custom_output_tuples)),     # Output(Signal, :semantic) => Id()
-                "output_tuples.register_additional_outputs"       => Normalizer.Task(OutputTuples.method(:register_additional_outputs)),     # Output(Signal, :semantic) => Id()
-                "output_tuples.filter_inherited_output_tuples"    => Normalizer.Task(OutputTuples.method(:filter_inherited_output_tuples)),
+                "output_tuples.normalize_output_tuples"           => OutputTuples.method(:normalize_output_tuples),     # Output(Signal, :semantic) => Id()
+                "output_tuples.remember_custom_output_tuples"     => OutputTuples.method(:remember_custom_output_tuples),     # Output(Signal, :semantic) => Id()
+                "output_tuples.register_additional_outputs"       => OutputTuples.method(:register_additional_outputs),     # Output(Signal, :semantic) => Id()
+                "output_tuples.filter_inherited_output_tuples"    => OutputTuples.method(:filter_inherited_output_tuples),
 # raise "move above Output code?"
             # Extension layer
-                "extensions.compute_normalizer_extensions" => Normalizer.Task(Extensions.method(:compute_normalizer_extensions)),
-                "extensions.compile_normalizer_extensions" => Normalizer.Task(Extensions.method(:compile_normalizer_extensions)),
+                "extensions.compute_normalizer_extensions" => Extensions.method(:compute_normalizer_extensions),
+                "extensions.compile_normalizer_extensions" => Extensions.method(:compile_normalizer_extensions),
 
                 # "step.normalize_task_wrap_extensions" => Normalizer.Task(TaskWrap.method(:normalize_task_wrap_extensions)),
                 # here, variable mapping is added.
-                "step.add_dsl_extensions_to_task_wrap_extensions" => Normalizer.Task(TaskWrap.method(:add_dsl_extensions_to_task_wrap_extensions)), # after this, we got a complete {:task_wrap_extensions} option.
-                "step.compile_task_wrap_from_extensions" => Normalizer.Task(TaskWrap.method(:compile_task_wrap_from_extensions)),
+                "step.add_dsl_extensions_to_task_wrap_extensions" => TaskWrap.method(:add_dsl_extensions_to_task_wrap_extensions), # after this, we got a complete {:task_wrap_extensions} option.
+                "step.compile_task_wrap_from_extensions" => TaskWrap.method(:compile_task_wrap_from_extensions),
 
-                "extensions.compile_recorded_extensions"  => Normalizer.Task(Extensions.method(:compile_recorded_extensions)), # DISCUSS: WHERE does this go?
+                "extensions.compile_recorded_extensions"  => Extensions.method(:compile_recorded_extensions), # DISCUSS: WHERE does this go?
 
-                "activity.wirings"                        => Normalizer.Task(OutputTuples::Connections.method(:compile_wirings)),
+                "activity.wirings"                        => OutputTuples::Connections.method(:compile_wirings),
 
 
 
                 # DISCUSS: make this configurable? maybe lots of folks don't want {:inherit}?
-                "inherit.compile_recorded_options"        => Normalizer.Task(Inherit.method(:compile_recorded_options)),
+                "inherit.compile_recorded_options"        => Inherit.method(:compile_recorded_options),
 
                 # TODO: make this a "Subprocess":
-                "activity.compile_data" => Normalizer.Task(method(:compile_data)),
-                "activity.create_row" => Normalizer.Task(method(:create_row)),
-                "activity.create_add" => Normalizer.Task(method(:create_add)),
-                "activity.create_adds" => Normalizer.Task(method(:create_adds)),
+                "activity.compile_data" => method(:compile_data),
+                "activity.create_row" => method(:create_row),
+                "activity.create_add" => method(:create_add),
+                "activity.create_adds" => method(:create_adds),
               }
             )
           end
@@ -165,16 +148,18 @@ module Trailblazer
           # DISCUSS: should we remove this special case?
           # This handles
           #   step task: :instance_method_exposing_circuit_interface
-          def macro_options_with_symbol_task(ctx, options:, **)
-            return if options[:wrap_task]
-            return unless options[:task].is_a?(Symbol)
+          def macro_options_with_symbol_task(ctx, flow_options, _, options:, **)
+            return ctx, flow_options if options[:wrap_task]
+            return ctx, flow_options unless options[:task].is_a?(Symbol)
 
-            ctx.merge(
+            ctx = ctx.merge(
               options: options.merge(
                 wrap_task:              true,
                 step_interface_builder: ->(task) { Trailblazer::Option(task) } # only wrap in Option, not {TaskAdapter}.
               )
             )
+
+            return ctx, flow_options
           end
 
           # @param {:options} The first argument passed to {#step}
@@ -182,55 +167,67 @@ module Trailblazer
           #
           # Specific to the "step DSL": if the first argument is a callable, wrap it in a {step_interface_builder}
           # since its interface expects the step interface, but the circuit will call it with circuit interface.
-          def normalize_step_interface(ctx, options:, **)
-            return if options.is_a?(Hash)
+          def normalize_step_interface(ctx, flow_options, _, options:, **)
+            return ctx, flow_options if options.is_a?(Hash)
 
             # Step Interface
             # step :find, ...
             # step Callable, ... (Method, Proc etc)
-            ctx.merge(
-              options: {
-                task:       options,
-                wrap_task:  true # task exposes step interface.
-              }
-            )
+            ctx = ctx.merge(
+                options: {
+                  task:       options,
+                  wrap_task:  true # task exposes step interface.
+                }
+              )
+
+            return ctx, flow_options
           end
 
           # @param :wrap_task If true, the {:task} is wrapped using the step_interface_builder, meaning the
           #                   task is expecting the step interface.
-          def wrap_task_with_step_interface(ctx, step_interface_builder:, task:, wrap_task: false, **)
-            return unless wrap_task
+          def wrap_task_with_step_interface(ctx, flow_options, _, step_interface_builder:, task:, wrap_task: false, **)
+            return ctx, flow_options unless wrap_task
 
-            ctx.merge(task: step_interface_builder.(task))
+            ctx = ctx.merge(task: step_interface_builder.(task))
+
+            return ctx, flow_options
           end
 
-          def normalize_id(ctx, task:, id: false, **)
-            ctx.merge(id: id || task)
+          def normalize_id(ctx, flow_options, _, task:, id: false, **)
+            ctx = ctx.merge(id: id || task)
+
+            return ctx, flow_options
           end
 
           # {:library_options} such as :sequence, :dsl_track, etc.
-          def merge_library_options(ctx, options:, library_options:, **)
-            ctx.merge(
-              options: library_options.merge(options)
-            )
+          def merge_library_options(ctx, flow_options, _, options:, library_options:, **)
+            ctx = ctx.merge(
+                options: library_options.merge(options)
+              )
+
+            return ctx, flow_options
           end
 
           # make ctx[:options] the actual ctx
-          def merge_user_options(ctx, options:, user_options:, **)
+          def merge_user_options(ctx, flow_options, _, options:, user_options:, **)
             # {options} are either a <#task> or {} from macro
-            ctx.merge(
+            ctx = ctx.merge(
               options: options.merge(user_options) # Note that the user options are merged over the macro options.
             )
+
+            return ctx, flow_options
           end
 
           # {:normalizer_options} such as {:track_name} get overridden by user/macro.
-          def merge_normalizer_options(ctx, normalizer_options:, options:, **)
-            ctx.merge(
+          def merge_normalizer_options(ctx, flow_options, _, normalizer_options:, options:, **)
+            ctx = ctx.merge(
               options: normalizer_options.merge(options)
             )
+
+            return ctx, flow_options
           end
 
-          def normalize_context(ctx, flow_options, _)
+          def normalize_context(ctx, flow_options, _, **)
             ctx = ctx[:options]
 
             return ctx, flow_options
@@ -238,16 +235,18 @@ module Trailblazer
 
           # Processes {:before,:after,:replace,:delete} options and
           # defaults to {before: "End.success"} which, yeah.
-          def normalize_sequence_insert(ctx, end_id:, **)
+          def normalize_sequence_insert(ctx, flow_options, _, end_id:, **)
             insertion = ctx.keys & sequence_insert_options.keys
             insertion = insertion[0]   || :before
             target    = ctx[insertion] || end_id
 
             insertion_option = sequence_insert_options[insertion]
 
-            ctx.merge(
+            ctx = ctx.merge(
               sequence_insert: {insertion_option => target}
             )
+
+            return ctx, flow_options
           end
 
           # Translate DSL option to "friendly interface" option.
@@ -261,11 +260,13 @@ module Trailblazer
             }
           end
 
-          def normalize_duplications(ctx, replace: false, **)
-            return if replace
+          def normalize_duplications(ctx, flow_options, _, replace: false, **)
+            return ctx, flow_options if replace
 
             raise_on_duplicate_id(ctx, **ctx.to_hash)
-            clone_duplicate_activity(ctx, **ctx.to_hash)
+            ctx = clone_duplicate_activity(ctx, **ctx.to_hash)
+
+            return ctx, flow_options
           end
 
           # @private
@@ -275,24 +276,26 @@ module Trailblazer
 
           # @private
           def clone_duplicate_activity(ctx, task:, sequence:, **)
-            return unless task.is_a?(Class)
-            return unless sequence.to_a.find { |_, row| row.task == task }
+            return ctx unless task.is_a?(Class)
+            return ctx unless sequence.to_a.find { |_, row| row.task == task }
 
             ctx.merge(task: task.clone)
           end
 
           # Whenever {:replace} and {:inherit} are passed, automatically assign {:id}.
           # DISCUSS: this step could be nested in {inherit_option}.
-          def id_with_inherit_and_replace(ctx, id: nil, replace: nil, inherit: nil, **)
-            return if id
-            return unless inherit # inherit: true and inherit: [] both work.
-            return unless replace
+          def id_with_inherit_and_replace(ctx, flow_options, _, id: nil, replace: nil, inherit: nil, **)
+            return ctx, flow_options if id
+            return ctx, flow_options unless inherit # inherit: true and inherit: [] both work.
+            return ctx, flow_options unless replace
 
-            ctx.merge(id: replace)
+            ctx = ctx.merge(id: replace)
+
+            return ctx, flow_options
           end
 
-          def create_row(ctx, task:, wirings:, magnetic_to:, data:, task_wrap:nil, **)
-            ctx.merge(
+          def create_row(ctx, flow_options, _, task:, wirings:, magnetic_to:, data:, task_wrap:nil, **)
+            ctx = ctx.merge(
               row: Sequence.Row(
                 task: task,
                 magnetic_to: magnetic_to,
@@ -301,36 +304,44 @@ module Trailblazer
                 task_wrap: task_wrap
               )
             )
+
+            return ctx, flow_options
           end
 
           # The {:add} field is one "friendly interface" instruction.
           # See {Adds.call} in the {activity} gem.
-          def create_add(ctx, row:, sequence_insert:, **)
-            ctx.merge(
+          def create_add(ctx, flow_options, _, row:, sequence_insert:, **)
+            ctx = ctx.merge(
               add: [
                 row,
                 id: row.id, # FIXME!
                 **sequence_insert
               ]
             )
+
+            return ctx, flow_options
           end
 
-          def create_adds(ctx, add:, adds:, **)
-            ctx.merge(
+          def create_adds(ctx, flow_options, _, add:, adds:, **)
+            ctx = ctx.merge(
               adds: [add] + adds
             )
+
+            return ctx, flow_options
           end
 
           # TODO: document DataVariable() => :name
           # Compile data that goes into the sequence row.
-          def compile_data(ctx, default_variables_for_data: [:id, :dsl_track], **)
+          def compile_data(ctx, flow_options, _, default_variables_for_data: [:id, :dsl_track], **)
             variables_for_data = ctx
               .find_all { |k, v| k.instance_of?(Linear::DataVariableName) }
               .flat_map { |k, v| Array(v) }
 
-            ctx.merge(
+            ctx = ctx.merge(
               data: (default_variables_for_data + variables_for_data).collect { |key| [key, ctx[key]] }.to_h
             )
+
+            return ctx, flow_options
           end